- Refactored some loops in MainProcess_dock - Moved print_received-method from MainProcess_dock to OpenEQuarterMain
